linezone_model_set/requirements.txt
cycler==0.10.0
Cython==0.29.2
dateutils==0.6.6
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.0
pandas==0.23.4
pbr==5.1.1
psycopg2-binary==2.7.6.1
py4j==0.10.7
pyparsing==2.3.1
pyspark==2.2.3
python-dateutil==2.7.5
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.2.16
stevedore==1.30.0
virtualenv==16.2.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
xlrd==1.2.0
